Refactored several components into their own file
