+++ renderer/pages/home.tsx
@@ -1,16 +1,18 @@
-import React from 'react';
-import Head from 'next/head';
+import { Alert, AlertTitle, Snackbar, styled } from '@mui/material';
+import { ipcRenderer, shell } from 'electron';
+
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
-import DialogTitle from '@mui/material/DialogTitle';
+import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
-import DialogActions from '@mui/material/DialogActions';
-import Typography from '@mui/material/Typography';
-import TextField from '@mui/material/TextField';
+import DialogTitle from '@mui/material/DialogTitle';
+import Head from 'next/head';
+import InitDialog from '../components/InitDialog';
 import Link from '../components/Link';
-import { Alert, AlertTitle, Snackbar, styled } from '@mui/material';
-import { shell, ipcRenderer } from 'electron';
+import React from 'react';
+import TextField from '@mui/material/TextField';
+import Typography from '@mui/material/Typography';
 import changePage from '../lib/page-transition';
 
 const Root = styled('div')(({ theme }) => {
@@ -23,15 +25,6 @@ const Root = styled('div')(({ theme }) => {
 
 function Home() {
     const [loginDialogOpen, setLoginDialogOpen] = React.useState(false);
-    const [initDialogOpen, setInitDialogOpen] = React.useState(false);
-
-    React.useEffect(() => {
-        ipcRenderer.send('mesg-db-get-registries', { table: 'user' });
-        ipcRenderer.once('reply-db-get-registries', (_, dbRows) => {
-            if (dbRows.length == 0)
-                setInitDialogOpen(true);
-        });
-    }, []);
 
     function LoginDialog() {
         let [username, setUsername] = React.useState('none');
@@ -96,73 +89,6 @@ function Home() {
                 <DialogActions>
                     <Button color="primary" onClick={handleLogin}>
                         Login
-                    </Button>
-                </DialogActions>
-            </Dialog>
-        );
-    }
-
-    function InitDialog() {
-        const [cleared, setCleared] = React.useState(false);
-        const fields = ["Email", "Password"];
-        const validations = React.useMemo(() => {
-            return {
-                email: false,
-                password: false
-            }
-        }, []);
-
-        const handleInit = () => {
-            const email = document.getElementById(fields[0].toLowerCase()) as HTMLInputElement;
-            const password = document.getElementById(fields[1].toLowerCase()) as HTMLInputElement;
-            ipcRenderer.send('mesg-db-create-user', { email: email.value, password: password.value });
-            ipcRenderer.once('reply-db-create-user', (_, success) => {
-                if (success) {
-                    setInitDialogOpen(false);
-                } else {
-                    // TODO Show error message
-                }
-            });
-        };
-
-        return (
-            <Dialog open={initDialogOpen}>
-                <DialogTitle>Setup DB</DialogTitle>
-                <DialogContent>
-                    <DialogContentText>
-                        Hi! Looks like this is your first time using DataWorks.
-                        Please enter the credentials for the admin user below.
-                    </DialogContentText>
-                    {fields.map((id) => {
-                        const capId = id;
-                        id = id.toLowerCase();
-                        const [error, setError] = React.useState(false);
-                        const validateContent = (event) => {
-                            const value = event.target.value;
-                            const invalid = !value || value.length == 0;
-                            setError(invalid);
-                            validations[id] = !invalid;
-                            setCleared(validations.email && validations.password);
-                        };
-                        return (
-                            <TextField
-                                autoFocus
-                                margin="dense"
-                                error={error}
-                                id={id}
-                                label={capId}
-                                key={id}
-                                type={id == "password" ? "password" : "text"}
-                                fullWidth
-                                variant="standard"
-                                onChange={validateContent}
-                            />
-                        );
-                    })}
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={handleInit} disabled={!cleared}>
-                        Continue
                     </Button>
                 </DialogActions>
             </Dialog>
